Added as new dataset and split the existing algorithm to work with both the new and the old datasets.
